Mandatory part is done

+++ ft_puthex_low.c
@@ -6,13 +6,13 @@
 /*   By: noaziki <noaziki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 11:44:27 by noaziki           #+#    #+#             */
-/*   Updated: 2024/11/27 13:32:04 by noaziki          ###   ########.fr       */
+/*   Updated: 2024/11/27 17:59:01 by noaziki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "printf.h"
+#include "ft_printf.h"
 
-int	ft_puthex_low(unsigned long n)
+int	ft_puthex_low(unsigned int n)
 {
 	int	counter;
