+++ ft_puthex_upp.c
@@ -6,13 +6,13 @@
 /*   By: noaziki <noaziki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 11:44:44 by noaziki           #+#    #+#             */
-/*   Updated: 2024/11/27 13:32:15 by noaziki          ###   ########.fr       */
+/*   Updated: 2024/11/27 17:59:08 by noaziki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "printf.h"
+#include "ft_printf.h"
 
-int	ft_puthex_upp(unsigned long n)
+int	ft_puthex_upp(unsigned int n)
 {
 	int	counter;
